Refactored folder names. Created components and store folder; and extermined 3 bugs by using destructure method.
