+++ client/components/app/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Navbar from '../features/navbar/Navbar';
+import {Navbar} from '../index';
 import AppRoutes from './AppRoutes';
 
 const App = () => {
