+++ client/components/app/AppRoutes.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Route, Routes } from 'react-router-dom';
-import AuthForm from '../features/auth/AuthForm';
-import Home from '../features/home/Home';
-import { me } from './store';
+import AuthForm from '../auth/AuthForm';
+import {Home} from "../index"
+import { me } from "../auth/authSlice"
 
 /**
  * COMPONENT
