+++ client/components/auth/AuthForm.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { authenticate } from '../../app/store';
+import { authenticate } from '../../store/store';
 
 /**
   The AuthForm component can be used for Login or Sign Up.
@@ -38,7 +38,6 @@ const AuthForm = ({ name, displayName }) => {
         <div>
           <button type="submit">{displayName}</button>
         </div>
-        {error && <div> {error} </div>}
       </form>
     </div>
   );
