+++ client/components/index.js
@@ -1,0 +1,3 @@
+export {default as Home} from './home/Home';
+export {default as Navbar} from './navbar/Navbar';
+export {default as AuthForm} from './auth/AuthForm';
