+++ client/components/navbar/Navbar.js
@@ -1,7 +1,11 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Link, useNavigate } from 'react-router-dom';
-import { logout } from '../../app/store';
+// import logout from store
+import { logout } from '../auth/authSlice';
+
+
+// import { logout } from '../';
 
 const Navbar = () => {
   const isLoggedIn = useSelector((state) => !!state.auth.me.id);
