+++ client/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import store from './app/store.js';
-import App from './app/App';
+import store from './store/store';
+import App from './components/app/App';
 import { BrowserRouter as Router } from 'react-router-dom';
 
 const root = createRoot(document.getElementById('app'));
